+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Constants.hs
@@ -7,7 +7,6 @@ module Test.Cardano.Ledger.Shelley.Constants (
 )
 where
 
-import Cardano.Ledger.BaseTypes (Version, natVersion)
 import Cardano.Ledger.Coin (Coin (..))
 import Data.Word (Word64)
 
@@ -93,8 +92,6 @@ data Constants = Constants
   , maxTreasury :: Integer
   , minReserves :: Integer
   , maxReserves :: Integer
-  , minMajorPV :: Version
-  , maxMajorPV :: Version
   , genTxStableUtxoSize :: Int
   -- ^ When generating Tx, we want the UTxO size to fluctuate around this point. If
   --   it gets too small, we can't balance the fee, too large it gets too complicated.
@@ -146,8 +143,6 @@ defaultConstants =
     , maxTreasury = 10000000
     , minReserves = 1000000
     , maxReserves = 10000000
-    , minMajorPV = natVersion @2
-    , maxMajorPV = maxBound
     , genTxStableUtxoSize = 100
     , genTxUtxoIncrement = 3
     }